Fix: fail to build in release configuration with Visual C++ 2008
git-svn-id: https://svn.sourceforge.jp/svnroot/magro/trunk@15 70dc4fbb-00e4-4569-b52f-2c4ca4cf82d4

+++ src/mainapp/win32/dirent.h
@@ -29,7 +29,7 @@ struct dirent *readdir(DIR *);
 void          rewinddir(DIR *);
 int mkdir(const char* pathname, int dummy);
 int chdir(const char* pathname);
-char *getcwd(char *buf, size_t size);
+char *getcwd(char *buf, int size);
 
 /*
